+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## 0.0.6-alpha (2025-05-03)
+
+### Changes
+
+- Added factories and seeders for roles and permissions
+- Command to publish seeders and factories to application
+- Configuration options for seeder behavior
 
 ## 0.0.5-alpha (2025-05-03)